I started writing IblCustomer and sorting some files

+++ BL/Location.cs
@@ -6,8 +6,8 @@ namespace IBL
     {
         public class Location
         {
-            public int Longitude { get; set; }
-            public int Lattitude { get; set; }
+            public double Longitude { get; set; }
+            public double Lattitude { get; set; }
 
         }
     }
